backend/events-manager/requirements.txt
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
jsonschema==4.20.0